chore: 기본 패키지 구조를 com.pickcar로 구성
