--- database/src/main/java/auth/domain/Auth.java
@@ -1,14 +1,0 @@
-package auth.domain;
-
-//FIXME: 토큰을 저장하는 테이블이 별도로 필요할까요 user에서 관리하는건?
-//@Entity
-//public class Auth {
-//
-//    @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-//    private Long id;
-//
-//    private String accessToken;
-//
-//    private String refreshToken;
-//}
