--- database/src/main/java/auth/domain/SuperAdmin.java
@@ -1,32 +1,0 @@
-package auth.domain;
-
-import global.domain.BaseEntity;
-import jakarta.persistence.Embedded;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@Table(name = "super_admins")
-@SuperBuilder(toBuilder = true)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class SuperAdmin extends BaseEntity {
-
-    @Embedded
-    private UserInfo userInfo;
-
-    @Enumerated(EnumType.STRING)
-    private UserRole userRole;
-
-    @Enumerated(EnumType.STRING)
-    private UserStatus status;
-
-}
