--- database/src/main/java/auth/domain/UserRole.java
@@ -1,5 +1,0 @@
-package auth.domain;
-
-public enum UserRole {
-    ADMIN, EMPLOYEE, SUPER_ADMIN
-}
