--- database/src/main/java/auth/domain/UserStatus.java
@@ -1,5 +1,0 @@
-package auth.domain;
-
-public enum UserStatus {
-    ACTIVE, INACTIVE
-}
