+++ database/src/main/java/com/pickcar/auth/domain/UserRole.java
@@ -1,0 +1,5 @@
+package com.pickcar.auth.domain;
+
+public enum UserRole {
+    ADMIN, EMPLOYEE, SUPER_ADMIN
+}
