+++ database/src/main/java/com/pickcar/auth/domain/UserStatus.java
@@ -1,0 +1,5 @@
+package com.pickcar.auth.domain;
+
+public enum UserStatus {
+    ACTIVE, INACTIVE
+}
