+++ database/src/main/java/com/pickcar/company/domain/ContractStatus.java
@@ -1,0 +1,5 @@
+package com.pickcar.company.domain;
+
+public enum ContractStatus {
+    ACTIVE, ENDED;
+}
