+++ database/src/main/java/com/pickcar/drivehistory/domain/DriveHistory.java
@@ -1,0 +1,38 @@
+package com.pickcar.drivehistory.domain;
+
+import com.pickcar.global.domain.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import java.time.LocalDateTime;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.SuperBuilder;
+import com.pickcar.reservation.domain.Reservation;
+
+@Entity
+@Getter
+@AllArgsConstructor
+@Table(name = "drive_histories")
+@SuperBuilder(toBuilder = true)
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class DriveHistory extends BaseEntity {
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "reservation_id", nullable = false)
+    private Reservation reservation;
+
+    @Column(nullable = false)
+    private LocalDateTime drivingStartedAt;
+
+    private LocalDateTime drivingEndedAt;
+
+    private Double totalDistance;
+
+    private LocalDateTime totalDrivingTime;
+}
