+++ database/src/main/java/com/pickcar/rental/domain/RentalStatus.java
@@ -1,0 +1,5 @@
+package com.pickcar.rental.domain;
+
+public enum RentalStatus {
+    RENTED, RETURNED
+}
