+++ database/src/main/java/com/pickcar/reservation/domain/ReservationStatus.java
@@ -1,0 +1,5 @@
+package com.pickcar.reservation.domain;
+
+public enum ReservationStatus {
+    RESERVED, CANCELLED, RETURNED;
+}
