--- database/src/main/java/company/domain/ContractStatus.java
@@ -1,5 +1,0 @@
-package company.domain;
-
-public enum ContractStatus {
-    ACTIVE, ENDED;
-}
