--- database/src/main/java/drivehistory/domain/DriveHistory.java
@@ -1,39 +1,0 @@
-package drivehistory.domain;
-
-import global.domain.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-import reservation.domain.Reservation;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@Table(name = "drive_histories")
-@SuperBuilder(toBuilder = true)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class DriveHistory extends BaseEntity {
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "reservation_id", nullable = false)
-    private Reservation reservation;
-
-    @Column(nullable = false)
-    private LocalDateTime drivingStartedAt;
-
-    private LocalDateTime drivingEndedAt;
-
-    private Double totalDistance;
-
-    private LocalDateTime totalDrivingTime;
-}
