--- database/src/main/java/rental/domain/Rental.java
@@ -1,45 +1,0 @@
-package rental.domain;
-
-import car.domain.Car;
-import company.domain.Company;
-import global.domain.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDateTime;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.SuperBuilder;
-
-@Entity
-@Getter
-@AllArgsConstructor
-@Table(name = "rentals")
-@SuperBuilder(toBuilder = true)
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Rental extends BaseEntity {
-
-    // 차와 회사는 연관관계로 설정 (지금은 단순 ID 사용)
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "car_id", nullable = false)
-    private Car car;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "company_id", nullable = false)
-    private Company company;
-
-    @Column(nullable = false)
-    private LocalDateTime rentedAt;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false)
-    private RentalStatus rentalStatus;
-
-}
