--- database/src/main/java/rental/domain/RentalStatus.java
@@ -1,5 +1,0 @@
-package rental.domain;
-
-public enum RentalStatus {
-    RENTED, RETURNED
-}
