--- database/src/main/java/reservation/domain/ReservationStatus.java
@@ -1,5 +1,0 @@
-package reservation.domain;
-
-public enum ReservationStatus {
-    RESERVED, CANCELLED, RETURNED;
-}
