--- domain/src/main/java/com/auth/infrastructure/SuperAdminRepository.java
@@ -1,7 +1,0 @@
-package com.auth.infrastructure;
-
-import auth.domain.SuperAdmin;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SuperAdminRepository extends JpaRepository<SuperAdmin, Long> {
-}
