--- domain/src/main/java/com/auth/infrastructure/UserRepository.java
@@ -1,7 +1,0 @@
-package com.auth.infrastructure;
-
-import auth.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
