--- domain/src/main/java/com/car/infrastructure/CarRepository.java
@@ -1,7 +1,0 @@
-package com.car.infrastructure;
-
-import car.domain.Car;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CarRepository extends JpaRepository<Car, Long> {
-}
