--- domain/src/main/java/com/company/infrastructure/CompanyRepository.java
@@ -1,7 +1,0 @@
-package com.company.infrastructure;
-
-import company.domain.Company;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-}
