--- domain/src/main/java/com/drivehistory/infrastructure/DriveHistoryRepository.java
@@ -1,7 +1,0 @@
-package com.drivehistory.infrastructure;
-
-import drivehistory.domain.DriveHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DriveHistoryRepository extends JpaRepository<DriveHistory, Long> {
-}
