+++ domain/src/main/java/com/pickcar/DomainApplication.java
@@ -1,0 +1,16 @@
+package com.pickcar;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+
+//FIXME: Application 실행 위치
+@EntityScan(basePackages = {
+        "**.domain"         //FIXME: 와일드 카드 경로 부적절 가능성 있음
+})
+@SpringBootApplication
+public class DomainApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(DomainApplication.class, args);
+    }
+}
