+++ domain/src/main/java/com/pickcar/auth/infrastructure/SuperAdminRepository.java
@@ -1,0 +1,7 @@
+package com.pickcar.auth.infrastructure;
+
+import com.pickcar.auth.domain.SuperAdmin;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SuperAdminRepository extends JpaRepository<SuperAdmin, Long> {
+}
