+++ domain/src/main/java/com/pickcar/auth/infrastructure/UserRepository.java
@@ -1,0 +1,7 @@
+package com.pickcar.auth.infrastructure;
+
+import com.pickcar.auth.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+}
