+++ domain/src/main/java/com/pickcar/company/infrastructure/CompanyRepository.java
@@ -1,0 +1,7 @@
+package com.pickcar.company.infrastructure;
+
+import com.pickcar.company.domain.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CompanyRepository extends JpaRepository<Company, Long> {
+}
