+++ domain/src/main/java/com/pickcar/drivehistory/infrastructure/DriveHistoryRepository.java
@@ -1,0 +1,7 @@
+package com.pickcar.drivehistory.infrastructure;
+
+import com.pickcar.drivehistory.domain.DriveHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DriveHistoryRepository extends JpaRepository<DriveHistory, Long> {
+}
