+++ domain/src/main/java/com/pickcar/reservation/infrastructure/ReservationRepository.java
@@ -1,0 +1,7 @@
+package com.pickcar.reservation.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import com.pickcar.reservation.domain.Reservation;
+
+public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+}
