--- domain/src/main/java/com/rental/infrastructure/RentalRepository.java
@@ -1,7 +1,0 @@
-package com.rental.infrastructure;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import rental.domain.Rental;
-
-public interface RentalRepository extends JpaRepository<Rental, Long> {
-}
