--- domain/src/main/java/com/reservation/infrastructure/ReservationRepository.java
@@ -1,7 +1,0 @@
-package com.reservation.infrastructure;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import reservation.domain.Reservation;
-
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-}
